+++ src/orbis2/database/orbis/entities/corpus_dao.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, VARCHAR, Sequence
+from sqlalchemy import Column, VARCHAR, Sequence, BigInteger
 from sqlalchemy.orm import relationship
 
 from orbis2.database.orbis.entities.corpus_supports_annotation_type_relation import \
@@ -9,7 +9,7 @@ from orbis2.database.orbis.orbis_base import OrbisBase
 class CorpusDao(OrbisBase):
     __tablename__ = 'corpus'
 
-    corpus_id = Column(Integer, Sequence('corpus_id_seq'), primary_key=True)
+    corpus_id = Column(BigInteger, Sequence('corpus_id_seq'), primary_key=True)
     name = Column(VARCHAR(40), nullable=False)
     runs = relationship('RunDao', back_populates='corpus')
     supported_annotation_types = relationship('AnnotationTypeDao', secondary=corpus_supports_annotation_type_table)
